Change losito to run on multiple MS files
