+++ data/noise/SEFD_LBA_INNER.csv
