+++ data/noise/SEFD_LBA_OUTER.csv
