+++ losito/operations/beam.py
@@ -13,12 +13,11 @@ def _run_parser(obs, parser, step):
     mode = parser.getstr( step, 'mode', 'default')
     usechannelfreq = parser.getbool( step, 'usechannelfreq', True)
     onebeamperpatch = parser.getbool( step, 'onebeamperpatch', False)
-
-    parser.checkSpelling( step, ['mode', 'usechannelfreq', 'onebeamperpatch'])
+    parser.checkSpelling( step, ['h5parmFilename', 'mode', 'usechannelfreq', 'onebeamperpatch'])
     return run(obs, mode, usechannelfreq, onebeamperpatch)
 
 
-def run(obs, mode='default', usechannelfreq=True, onebeamperpatch=False):
+def run(obs,   mode='default', usechannelfreq=True, onebeamperpatch=False):
     """
     Corrupts with the beam model.
 
@@ -39,9 +38,10 @@ def run(obs, mode='default', usechannelfreq=True, onebeamperpatch=False):
         return 1
 
     # Update predict parset parameters for the obs
-    obs.parset_parameters['predict.usebeammodel'] = True
-    obs.parset_parameters['predict.beammode'] = mode
-    obs.parset_parameters['predict.usechannelfreq'] = usechannelfreq
-    obs.parset_parameters['predict.onebeamperpatch'] = onebeamperpatch
+    for ms in obs:
+        ms.parset_parameters['predict.usebeammodel'] = True
+        ms.parset_parameters['predict.beammode'] = mode
+        ms.parset_parameters['predict.usechannelfreq'] = usechannelfreq
+        ms.parset_parameters['predict.onebeamperpatch'] = onebeamperpatch
 
     return 0
