--- losito/progress.py
@@ -1,40 +1,0 @@
-# The MIT License (MIT)
-# Copyright (c) 2016 Vladimir Ignatev
-#
-# Permission is hereby granted, free of charge, to any person obtaining
-# a copy of this software and associated documentation files (the "Software"),
-# to deal in the Software without restriction, including without limitation
-# the rights to use, copy, modify, merge, publish, distribute, sublicense,
-# and/or sell copies of the Software, and to permit persons to whom the Software
-# is furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included
-# in all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
-# INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
-# PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE
-# FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT
-# OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE
-# OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-import sys
-
-'''
-Use this simple script as progressbar for now, since the old progressbar
-library is python 2 and not supported anymore. In the future, it might be
-usefull to use the tqdm library for this: https://github.com/tqdm/tqdm 
-'''
-def progress(count, total, status=''):
-    '''Usage: place in for-loop like:
-        for i, val in enumerate(vals):
-            progress(i, len(vals), somestringcomment)
-            ...
-    '''
-    bar_len = 40
-    filled_len = round(bar_len * count / float(total))
-
-    percents = round(100.0 * count / float(total), 1)
-    bar = '=' * filled_len + '-' * (bar_len - filled_len)
-
-    sys.stdout.write('[%s] %s%s ...%s\r' % (bar, percents, '%', status))   
-    sys.stdout.flush()
